# docs/openapi.yaml
openapi: 3.0.3
info:
  title: nox-mem API
  description: |
    Pain-weighted hybrid memory API for LLM agents — yours by design.

    This spec covers all HTTP endpoints across Wave A through D implemented in
    nox-mem. Some endpoints are GATED by env vars and return 503 when not
    enabled. All endpoints listen on port 18802 by default (NOX_API_PORT env
    var), never 18800 (Chrome squats that port).

    ## Authentication

    Authentication is optional by default. If `NOX_API_TOKEN` is set in the
    server environment, callers must include `Authorization: Bearer <token>` on
    all requests. Without the env var the server allows all localhost callers.
    The `/api/export` and `/api/import` endpoints always require auth when any
    token is configured (same mechanism — `requireApiToken()` middleware applied
    in front of the handler).

    ## Gated features

    The following env vars gate features. When the var is absent/disabled, the
    affected endpoints return `503 Service Unavailable` with `{"error":
    "feature disabled", "env_var": "<NOX_X>"}`:

    | Feature | Env var | Endpoints |
    |---------|---------|-----------|
    | Answer primitive (P1) | `NOX_ANSWER_ENABLED=1` | `POST /api/answer` |
    | Hooks autocapture (P2) | `NOX_HOOKS_ENABLED=1` | `/api/hooks/*` |
    | Realtime viewer (P5) | `NOX_VIEWER_ENABLED=1` | `/api/events/stream`, `/viewer/*` |
    | Conflict detection (L2) | `NOX_KG_CONFLICTS_ENABLED=1` | `/api/kg/conflicts*` |
...

...

...

...
  contact:
...
servers:
...
security:
  - {}
  - BearerAuth: []
...
paths:
...
  /api/search:
    get:
      tags: [Search]
      summary: Hybrid search over memory chunks
      description: |
        3-layer hybrid search: FTS5 BM25 → Gemini semantic embedding →
        RRF fusion (k=60). Accepts the query as the `q` parameter.

        **P3 additions:** `as_of` and `changed_since` accept flexible date
        expressions (ISO 8601 or relative like `7d`, `2w`).
      operationId: searchGet
      parameters:
        - name: q
          in: query
          required: true
          description: Search query string
          schema:
            type: string
            minLength: 1
          example: "Gemini quota exceeded nightly cron"
        - name: limit
          in: query
          required: false
          description: Max results to return. Default 10.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: as_of
          in: query
          required: false
          description: |
            Return only chunks whose `created_at` is on or before this date.
            Accepts ISO 8601 (`2026-05-01`) or relative expressions (`7d`, `2w`).
          schema:
            type: string
          example: "7d"
        - name: changed_since
          in: query
          required: false
          description: |
            Return only chunks updated after this date. Same format as `as_of`.
          schema:
            type: string
          example: "2026-05-01"
      responses:
        "200":
          description: Ranked list of matching chunks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
              example:
                - chunk_id: 41203
                  content: "Gemini 2.5 Flash Lite is the default model for all agent infra tasks..."
                  score: 0.913
                  source_path: "memory/entities/decision/model-selection.md"
                  section: compiled
                  chunk_type: decision
                  created_at: "2026-04-22T14:30:00Z"
        "400":
          description: Missing or invalid query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "q parameter required"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Search]
      summary: Hybrid search (POST body variant)
      description: |
        Same as `GET /api/search` but accepts parameters in a JSON body.
        Useful when the query string is long or when `as_of`/`changed_since`
        are passed programmatically.
      operationId: searchPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
            example:
              q: "how to reapply monkey-patch after openclaw upgrade"
              limit: 8
              as_of: "2026-05-10"
      responses:
        "200":
          description: Ranked list of matching chunks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/kg:
    get:
      tags: [Knowledge Graph]
      summary: Knowledge graph — entities and relations
      description: |
        Returns up to 200 most-mentioned entities and up to 500 most-confident
        relations. Intended for dashboard visualization. Note: `source` and
        `target` are canonical entity names (strings), not IDs — per the DB
        JOIN in `api-server.ts`.
      operationId: getKg
      responses:
        "200":
          description: KG snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KgResponse"
              example:
                entities:
                  - id: 12
                    name: "openclaw-gateway"
                    type: service
                    mentions: 847
                relations:
                  - source: "openclaw-gateway"
                    relation: "depends_on"
                    target: "nox-mem-api"
                    confidence: 0.92
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/export:
    post:
      tags: [Export/Import (A2)]
      summary: Export memory to a portable archive
      description: |
        Streams a gzip tar archive containing the memory corpus. The archive
        includes `manifest.json`, `chunks.jsonl`, `embeddings.bin`,
        `embeddings.idx`, `kg_entities.jsonl`, `kg_relations.jsonl`, and
        `ops_audit.jsonl`. Files are optionally encrypted with AES-256-GCM
        (scrypt KDF).

        **Gating:** requires `NOX_ARCHIVE_ENABLED=1`.

        **Auth:** always requires `Authorization: Bearer <token>` when
        `NOX_API_TOKEN` is set.

        **Streaming:** response is `Transfer-Encoding: chunked`. For large
        corpora (~1.5 GB), do not buffer — pipe directly to disk or cloud
        storage.

        **Response headers:**
        - `X-Nox-Export-Manifest-Checksum`: sha256 hex of manifest file
        - `X-Nox-Export-Chunks-Count`: number of exported chunks
        - `X-Nox-Export-Schema-Version`: DB schema version at export time

        **Implementation file:** `staged/A2/edits/src/lib/archive/`

        **Spec:** `specs/2026-05-17-A2-export-import.md §4`
      operationId: postExport
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
            example:
              project: "granix"
              since: "2026-01-01"
              format: "tar"
              exclude_embeddings: false
              encrypt: false
      responses:
        "200":
          description: Streaming archive (gzip tar or encrypted binary)
          headers:
            X-Nox-Export-Manifest-Checksum:
              schema:
                type: string
                description: sha256 hex of the manifest file
            X-Nox-Export-Chunks-Count:
              schema:
                type: integer
            X-Nox-Export-Schema-Version:
              schema:
                type: integer
          content:
            application/gzip:
              schema:
                type: string
                format: binary
                description: gzip tar archive
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: AES-256-GCM encrypted archive (when encrypt=true)
        "400":
          description: Invalid filter parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/FeatureDisabled"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/chunk/{id}/mark:
    post:
      tags: [Confidence (L3)]
      summary: Mark a chunk as canonical, refuted, or stale
      description: |
        Sets the `provenance_kind` to `user-marked` on the specified chunk and
        applies the corresponding confidence value:
        - `canonical` → confidence = `NOX_CONFIDENCE_USER_CANONICAL` (default 0.95)
        - `refuted` → confidence = `NOX_CONFIDENCE_USER_REFUTED` (default 0.05)
        - `stale` → confidence = 0.3, no provenance change

        The operation is recorded in `ops_audit`. Returns the applied
        confidence and provenance, and the audit id.

        **Gating:** always available; confidence values only affect search
        ranking when `NOX_RANKING_CONFIDENCE=active`.

        **Implementation file:** `staged/L3/edits/src/api/mark.ts`
      operationId: markChunk
      parameters:
        - name: id
          in: path
          required: true
          description: Chunk id
          schema:
            type: integer
          example: 41203
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkRequest"
            example:
              kind: canonical
              notes: "Verified this decision is still in effect as of 2026-05-18"
      responses:
        "200":
          description: Mark applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkResult"
              example:
                ok: true
                chunk_id: 41203
                applied:
                  confidence: 0.95
                  provenance_kind: user-marked
                audit_id: 1047
        "400":
          description: Invalid chunk id or mark kind
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkError"
              examples:
                bad_id:
                  value:
                    ok: false
                    error: "invalid chunk id: abc"
                    code: bad_id
                bad_kind:
                  value:
                    ok: false
                    error: "invalid kind: archived — expected canonical|refuted|stale"
                    code: bad_kind
        "404":
          description: Chunk not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkError"
              example:
                ok: false
                error: "chunk 99999 not found"
                code: not_found
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/hooks/status:
    get:
      tags: [Hooks (P2)]
      summary: Hooks pipeline configuration and queue depth
      description: |
        Returns the current hooks pipeline configuration (from
        `loadConfig()`) and live queue metrics. Never returns raw event
        payloads.

        **Gating:** requires `NOX_HOOKS_ENABLED=1`.

        **Implementation file:**
        `.claude/worktrees/agent-a21b345bc854329aa/staged/P2/edits/src/api/hooks.ts`
      operationId: getHooksStatus
      responses:
        "200":
          description: Hooks status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HooksStatus"
              example:
                config:
                  enabled: true
                  allowed_sources: [mcp, api, cli, slack]
                  rate_limit_per_min: 60
                  dedup_threshold: 0.85
                  llm_classify: false
                  dry_run: false
                  queue_size: 100
                  min_length: 20
                  pii_policy: redact
                queueDepth: 3
                rateLimitTokens: 47
        "503":
          $ref: "#/components/responses/FeatureDisabled"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        `NOX_API_TOKEN` — set in server `.env`. Required on all routes when
        the var is present. For the viewer endpoints specifically,
        `NOX_VIEWER_AUTH_TOKEN` may be configured separately. Without any
        token configured, the server allows all calls (localhost-only by
        design).
...